Add same-letter cycling for file-jumping
This will allow files selected by their
first letter to be cycled when the same
letter is pressed.

e.g. having files "a.txt" and "ab.txt"
Pressing "a" the first time will select
"a.txt", pressing "a" a second time will
selected "ab.txt" instead of reselecting
"a.txt"