Added the loading indicator, file icons, onedrive status...
I Got the File icons to load, added the OneDrive Sync Status, also added the Column Layout menuitem to the statusbar. Also added the loading indicator functionality when going through folders. Right click now works.

I also made a change so that the current folder selected is blue but when a subfolder is selected then it is greyed out