Update Resources.resw
Hi. I'm Bruno. I am a young programmer who likes to help incredible communities and projects. I come to propose a change in this pt-PT file. I did this because I noticed several spelling mistakes and translation into Portuguese from Portugal and, as a fluent Portuguese from Portugal, I decided to propose this change. I hope it is accepted and that it can contribute to better the Files application in Portuguese of Portugal.

Compliments

Esibruti