Replace ComboBox with custom ComboBoxEx control
Implemented a new custom control, `ComboBoxEx`, which extends the `ComboBox` control and overrides `OnDropDownOpened`, `OnDropDownClosed`, and `MeasureOverride` methods to adjust and cache the control's width. Added a new namespace `uc` to reference the `Files.App.UserControls` namespace in various XAML files. Also, introduced a new file `ComboBoxExtensions.cs` that provides an attached property `IsKeepWidthEnabled` to adjust the width of the `ComboBox` when the dropdown is opened and closed.