`Remove ComboBoxExtensions.cs file`
` `

`Removed the entire ComboBoxExtensions.cs file which contained the ComboBoxExtensions class. This class provided an extension for the ComboBox control in the WinUI library, controlling the width of the ComboBox dropdown. It had a dependency property IsKeepWidthEnabled that maintained the dropdown width as it was when first opened by handling the DropDownOpened and DropDownClosed events. This change eliminates this functionality.`