Fix: Optimize the code logic for loading the file list
There's no need to perform reordering every time some files are loaded, because the interval between each reordering is 500ms, which is too short for users to see the sorted results before it's reordered again. This means that during the file loading process, the sorting results are useless to users, and it also causes a lot of extra CPU computational overhead, leading to lag in the file loading process. It's sufficient to perform a single sorting of all files only after the file list has finished loading completely.

Signed-off-by: Tptogiar <Tptogiar@gmail.com>