AddMyDriveToPathAndValidate; is_my_drive
Renamed `ValidatePath` to `AddMyDriveToPathAndValidate`, and made it also append
`My Drive' to the path it is given.

The only path we ever get from the media table is the drive letter (if `Google
Drive streaming location` is set to `Drive letter`, otherwise we get nothing
from the media table) followed by `:\`. As such, I added a call to
`AddMyDriveToPathAndValidate' on this path.

We only find paths in the roots table whose `is_my_drive` is true when `My Drive
syncing options` is set to Mirror files. I believe these paths are the ones Josh
was referring to when he said

> Google Drive set to mirror to a user selected folder will show both the
> redirect drive/folder (This show one My Drive folder) and another that goes
> directly to the store folder location.

The path we get from roots is the one that "goes directly to the store folder
location" (hence its `is_my_drive` is true).

With this PR, we now get the "redirect drive/folder" as well, from the registry.
As such, since Josh said

> Preferably it would only show the redirect drive/folder and add /My Drive to
> the path, then hide all other locations

I added a call to `AddMyDriveToPathAndValidate` on every path we get from the
registry, and I made the roots table read loop `continue` if it finds a path
whose `is_my_drive` is true.