fix issue where Google Drive appears under drives
Add Debug.WriteLine logging to track when App.AppModel.GoogleDrivePath is
written and read.

Add logic to RemoteDrivesService.GetDrivesAsync() that filters out Google Drive
by calling GoogleDriveCloudDetector.GetGoogleDriveProvidersFromRegistryAsync().

Adjust GoogleDriveCloudDetector.GetGoogleDriveProvidersFromRegistryAsync() so it
can be called from RemoteDrivesService.GetDrivesAsync(), and so it can be told
to not append 'My Drive'.