Updated StorageRing
I changed the interpolation to be based on percentage rather than value between min and max - this is consistent with the StorageBar's interpolation at the start and end.

Storage Ring's track interpolation at the end is still done with value to angle conversions as the Angle is the best way to know when the start and end angles meet.