Changed "Percent" accessibility and rounding
I have made Percent a non-protected DependencyProperty so you can do binding to it.  That way any text label you wish to display the Percentage, will be able to bind to it.

I have made it so changing the value externally does nothing, and it will update it again as the value changes.

Also added a Percent TextBlock to the examples on the Testing Page